Refactor Object Detection and PCS pages with new components and metadata updates
- Replaced the main content of the Object Detection and PCS pages with dedicated components (ObjectDetectionClient and PCSClient) for better modularity.
- Updated metadata for both pages to enhance SEO and provide clearer descriptions.
- Added new model entries in the PCS sample results JSON files, including relevant metadata and performance metrics.
- Adjusted the visibility of certain columns and filters to improve user experience and data presentation.
